pyproject.toml
[project]
name = "ela-vector-search"
version = "0.1.0"
description = "Vector search for ELA journal articles using Qdrant and LangChain"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.5",
    "langchain-core>=0.1.15",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.16",
    "qdrant-client>=1.7.0",
    "pdfplumber>=0.10.3",
    "pypdf>=4.0.0",
    "langgraph>=0.0.19",
    "langchain-qdrant>=0.0.2",
    "ragas>=0.1.0",
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0"
]
requires-python = ">=3.9,<4.0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true